added tests for GET
